Updated FloatingPointValue to use float not double for its internal representation.
git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006730 13f79535-47bb-0310-9956-ffa450edef68

+++ no-transactions/test/org/nakedobjects/object/transaction/Account.java
@@ -36,7 +36,7 @@ public class Account {
     }
 
     public boolean canWithdraw(Transfer transfer) {
-        return getBalance().doubleValue() - transfer.getAmount().doubleValue() >= 0.0;
+        return getBalance().floatValue() - transfer.getAmount().floatValue() >= 0.0;
     }
 
     public void deposit(Transfer transfer) {
